Style cleanup + Started on named warps

+++ src/main/java/com/bendude56/goldenapple/antigrief/BlockTNT.java
@@ -80,7 +80,7 @@ public class BlockTNT extends net.minecraft.server.v1_6_R2.BlockTNT {
 	@Override
 	public void wasExploded(World world, int i, int j, int k, Explosion explosion) {
 		if (!world.isStatic && !GoldenApple.getInstanceMainConfig().getBoolean("modules.antigrief.noExplosionTnt", true)) {
-            EntityTNTPrimed entitytntprimed = new EntityTNTPrimed(world, (double) ((float) i + 0.5F), (double) ((float) j + 0.5F), (double) ((float) k + 0.5F), explosion.c());
+            EntityTNTPrimed entitytntprimed = new EntityTNTPrimed(world, i + 0.5F, j + 0.5F, k + 0.5F, explosion.c());
 
             entitytntprimed.fuseTicks = world.random.nextInt(entitytntprimed.fuseTicks / 4) + entitytntprimed.fuseTicks / 8;
             world.addEntity(entitytntprimed);
